feat: Change time units to seconds and stroke units to mm in dashboard
UI Display Improvements:
- Change heating/cooling time display from milliseconds to seconds with 2 decimal precision
- Convert stroke display from μm to mm (divide by 1000) for better readability
- Update CycleData dataclass: heating_time/cooling_time changed from int to float
- Update column headers: "Heat(ms)" → "Heat(s)", "Cool(ms)" → "Cool(s)"

Data Conversion Changes:
- main_window.py: Convert timing from ms to seconds (round to 2 decimals)
- results_table_widget.py: Display stroke as mm by dividing by 1000
- gui_state_manager.py: Change CycleData time fields from int to float

Display Format Examples:
- Before: Heat: 200ms, Cool: 300ms, Stroke: 170000
- After: Heat: 0.20s, Cool: 0.30s, Stroke: 170.0

+++ src/ui/gui/services/gui_state_manager.py
@@ -27,8 +27,8 @@ class CycleData:
     temperature: float  # Temperature in °C
     stroke: float  # Stroke position in mm
     force: float  # Measured force in kgf
-    heating_time: int  # Heating time in ms
-    cooling_time: int  # Cooling time in ms
+    heating_time: float  # Heating time in seconds
+    cooling_time: float  # Cooling time in seconds
     status: str  # Cycle status (PASS/FAIL)
